+++ src/wallet/serviceWorker/utils.ts
@@ -1,5 +1,3 @@
-import { ExtendedVirtualCoin, VirtualCoin, Wallet } from "../..";
-
 /**
  * setupServiceWorker sets up the service worker.
  * @param path - the path to the service worker script
@@ -53,16 +51,4 @@ export async function setupServiceWorker(path: string): Promise<ServiceWorker> {
         navigator.serviceWorker.addEventListener("activate", onActivate);
         navigator.serviceWorker.addEventListener("error", onError);
     });
-}
-
-export function extendVirtualCoin(
-    wallet: Wallet,
-    vtxo: VirtualCoin
-): ExtendedVirtualCoin {
-    return {
-        ...vtxo,
-        forfeitTapLeafScript: wallet.offchainTapscript.forfeit(),
-        intentTapLeafScript: wallet.offchainTapscript.exit(),
-        tapTree: wallet.offchainTapscript.encode(),
-    };
 }
