+++ src/wallet/utils.ts
@@ -1,0 +1,13 @@
+import { ExtendedVirtualCoin, VirtualCoin, Wallet } from "..";
+
+export function extendVirtualCoin(
+    wallet: Wallet,
+    vtxo: VirtualCoin
+): ExtendedVirtualCoin {
+    return {
+        ...vtxo,
+        forfeitTapLeafScript: wallet.offchainTapscript.forfeit(),
+        intentTapLeafScript: wallet.offchainTapscript.exit(),
+        tapTree: wallet.offchainTapscript.encode(),
+    };
+}
